refactor: use biome instead of prettier & eslint
