+++ nodejs/__test__/tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist/spec",
-    "module": "commonjs",
-    "target": "es2022",
-    "types": ["jest", "node"]
-  },
-  "include": ["**/*"]
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./dist/spec",
+		"module": "commonjs",
+		"target": "es2022",
+		"types": ["jest", "node"]
+	},
+	"include": ["**/*"]
 }
