+++ nodejs/eslint.config.js
@@ -7,22 +7,22 @@ const eslintConfigPrettier = require("eslint-config-prettier");
 const jsdoc = require("eslint-plugin-jsdoc");
 
 module.exports = tseslint.config(
-  eslint.configs.recommended,
-  jsdoc.configs["flat/recommended"],
-  eslintConfigPrettier,
-  ...tseslint.configs.recommended,
-  {
-    rules: {
-      "@typescript-eslint/naming-convention": "error",
-      "jsdoc/require-returns": "off",
-      "jsdoc/require-param": "off",
-      "jsdoc/require-jsdoc": [
-        "error",
-        {
-          publicOnly: true,
-        },
-      ],
-    },
-    plugins: jsdoc,
-  },
+	eslint.configs.recommended,
+	jsdoc.configs["flat/recommended"],
+	eslintConfigPrettier,
+	...tseslint.configs.recommended,
+	{
+		rules: {
+			"@typescript-eslint/naming-convention": "error",
+			"jsdoc/require-returns": "off",
+			"jsdoc/require-param": "off",
+			"jsdoc/require-jsdoc": [
+				"error",
+				{
+					publicOnly: true,
+				},
+			],
+		},
+		plugins: jsdoc,
+	},
 );
