+++ nodejs/jest.config.js
@@ -1,7 +1,7 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  moduleDirectories: ["node_modules", "./dist"],
-  moduleFileExtensions: ["js", "ts"],
+	preset: "ts-jest",
+	testEnvironment: "node",
+	moduleDirectories: ["node_modules", "./dist"],
+	moduleFileExtensions: ["js", "ts"],
 };
