+++ nodejs/lancedb/arrow.ts
@@ -13,25 +13,25 @@
 // limitations under the License.
 
 import {
-  Field,
-  makeBuilder,
-  RecordBatchFileWriter,
-  Utf8,
-  type Vector,
-  FixedSizeList,
-  vectorFromArray,
-  Schema,
-  Table as ArrowTable,
-  RecordBatchStreamWriter,
-  List,
-  RecordBatch,
-  makeData,
-  Struct,
-  type Float,
-  DataType,
-  Binary,
-  Float32,
-  type makeTable,
+	Table as ArrowTable,
+	Binary,
+	DataType,
+	Field,
+	FixedSizeList,
+	type Float,
+	Float32,
+	List,
+	RecordBatch,
+	RecordBatchFileWriter,
+	RecordBatchStreamWriter,
+	Schema,
+	Struct,
+	Utf8,
+	type Vector,
+	makeBuilder,
+	makeData,
+	type makeTable,
+	vectorFromArray,
 } from "apache-arrow";
 import { type EmbeddingFunction } from "./embedding/embedding_function";
 import { sanitizeSchema } from "./sanitize";
@@ -43,64 +43,64 @@ export type Data = Record<string, unknown>[] | ArrowTable;
  * Options to control how a column should be converted to a vector array
  */
 export class VectorColumnOptions {
-  /** Vector column type. */
-  type: Float = new Float32();
+	/** Vector column type. */
+	type: Float = new Float32();
 
-  constructor(values?: Partial<VectorColumnOptions>) {
-    Object.assign(this, values);
-  }
+	constructor(values?: Partial<VectorColumnOptions>) {
+		Object.assign(this, values);
+	}
 }
 
 /** Options to control the makeArrowTable call. */
 export class MakeArrowTableOptions {
-  /*
-   * Schema of the data.
-   *
-   * If this is not provided then the data type will be inferred from the
-   * JS type.  Integer numbers will become int64, floating point numbers
-   * will become float64 and arrays will become variable sized lists with
-   * the data type inferred from the first element in the array.
-   *
-   * The schema must be specified if there are no records (e.g. to make
-   * an empty table)
-   */
-  schema?: Schema;
+	/*
+	 * Schema of the data.
+	 *
+	 * If this is not provided then the data type will be inferred from the
+	 * JS type.  Integer numbers will become int64, floating point numbers
+	 * will become float64 and arrays will become variable sized lists with
+	 * the data type inferred from the first element in the array.
+	 *
+	 * The schema must be specified if there are no records (e.g. to make
+	 * an empty table)
+	 */
+	schema?: Schema;
 
-  /*
-   * Mapping from vector column name to expected type
-   *
-   * Lance expects vector columns to be fixed size list arrays (i.e. tensors)
-   * However, `makeArrowTable` will not infer this by default (it creates
-   * variable size list arrays).  This field can be used to indicate that a column
-   * should be treated as a vector column and converted to a fixed size list.
-   *
-   * The keys should be the names of the vector columns.  The value specifies the
-   * expected data type of the vector columns.
-   *
-   * If `schema` is provided then this field is ignored.
-   *
-   * By default, the column named "vector" will be assumed to be a float32
-   * vector column.
-   */
-  vectorColumns: Record<string, VectorColumnOptions> = {
-    vector: new VectorColumnOptions(),
-  };
-  embeddings?: EmbeddingFunction<unknown>;
+	/*
+	 * Mapping from vector column name to expected type
+	 *
+	 * Lance expects vector columns to be fixed size list arrays (i.e. tensors)
+	 * However, `makeArrowTable` will not infer this by default (it creates
+	 * variable size list arrays).  This field can be used to indicate that a column
+	 * should be treated as a vector column and converted to a fixed size list.
+	 *
+	 * The keys should be the names of the vector columns.  The value specifies the
+	 * expected data type of the vector columns.
+	 *
+	 * If `schema` is provided then this field is ignored.
+	 *
+	 * By default, the column named "vector" will be assumed to be a float32
+	 * vector column.
+	 */
+	vectorColumns: Record<string, VectorColumnOptions> = {
+		vector: new VectorColumnOptions(),
+	};
+	embeddings?: EmbeddingFunction<unknown>;
 
-  /**
-   * If true then string columns will be encoded with dictionary encoding
-   *
-   * Set this to true if your string columns tend to repeat the same values
-   * often.  For more precise control use the `schema` property to specify the
-   * data type for individual columns.
-   *
-   * If `schema` is provided then this property is ignored.
-   */
-  dictionaryEncodeStrings: boolean = false;
+	/**
+	 * If true then string columns will be encoded with dictionary encoding
+	 *
+	 * Set this to true if your string columns tend to repeat the same values
+	 * often.  For more precise control use the `schema` property to specify the
+	 * data type for individual columns.
+	 *
+	 * If `schema` is provided then this property is ignored.
+	 */
+	dictionaryEncodeStrings: boolean = false;
 
-  constructor(values?: Partial<MakeArrowTableOptions>) {
-    Object.assign(this, values);
-  }
+	constructor(values?: Partial<MakeArrowTableOptions>) {
+		Object.assign(this, values);
+	}
 }
 
 /**
@@ -196,114 +196,114 @@ export class MakeArrowTableOptions {
  * ```
  */
 export function makeArrowTable(
-  data: Array<Record<string, unknown>>,
-  options?: Partial<MakeArrowTableOptions>,
+	data: Array<Record<string, unknown>>,
+	options?: Partial<MakeArrowTableOptions>,
 ): ArrowTable {
-  if (
-    data.length === 0 &&
-    (options?.schema === undefined || options?.schema === null)
-  ) {
-    throw new Error("At least one record or a schema needs to be provided");
-  }
+	if (
+		data.length === 0 &&
+		(options?.schema === undefined || options?.schema === null)
+	) {
+		throw new Error("At least one record or a schema needs to be provided");
+	}
 
-  const opt = new MakeArrowTableOptions(options !== undefined ? options : {});
-  if (opt.schema !== undefined && opt.schema !== null) {
-    opt.schema = sanitizeSchema(opt.schema);
-    opt.schema = validateSchemaEmbeddings(opt.schema, data, opt.embeddings);
-  }
-  const columns: Record<string, Vector> = {};
-  // TODO: sample dataset to find missing columns
-  // Prefer the field ordering of the schema, if present
-  const columnNames =
-    opt.schema != null ? (opt.schema.names as string[]) : Object.keys(data[0]);
-  for (const colName of columnNames) {
-    if (
-      data.length !== 0 &&
-      !Object.prototype.hasOwnProperty.call(data[0], colName)
-    ) {
-      // The field is present in the schema, but not in the data, skip it
-      continue;
-    }
-    // Extract a single column from the records (transpose from row-major to col-major)
-    let values = data.map((datum) => datum[colName]);
+	const opt = new MakeArrowTableOptions(options !== undefined ? options : {});
+	if (opt.schema !== undefined && opt.schema !== null) {
+		opt.schema = sanitizeSchema(opt.schema);
+		opt.schema = validateSchemaEmbeddings(opt.schema, data, opt.embeddings);
+	}
+	const columns: Record<string, Vector> = {};
+	// TODO: sample dataset to find missing columns
+	// Prefer the field ordering of the schema, if present
+	const columnNames =
+		opt.schema != null ? (opt.schema.names as string[]) : Object.keys(data[0]);
+	for (const colName of columnNames) {
+		if (
+			data.length !== 0 &&
+			!Object.prototype.hasOwnProperty.call(data[0], colName)
+		) {
+			// The field is present in the schema, but not in the data, skip it
+			continue;
+		}
+		// Extract a single column from the records (transpose from row-major to col-major)
+		let values = data.map((datum) => datum[colName]);
 
-    // By default (type === undefined) arrow will infer the type from the JS type
-    let type;
-    if (opt.schema !== undefined) {
-      // If there is a schema provided, then use that for the type instead
-      type = opt.schema?.fields.filter((f) => f.name === colName)[0]?.type;
-      if (DataType.isInt(type) && type.bitWidth === 64) {
-        // wrap in BigInt to avoid bug: https://github.com/apache/arrow/issues/40051
-        values = values.map((v) => {
-          if (v === null) {
-            return v;
-          }
-          if (typeof v === "bigint") {
-            return v;
-          }
-          if (typeof v === "number") {
-            return BigInt(v);
-          }
-          throw new Error(
-            `Expected BigInt or number for column ${colName}, got ${typeof v}`,
-          );
-        });
-      }
-    } else {
-      // Otherwise, check to see if this column is one of the vector columns
-      // defined by opt.vectorColumns and, if so, use the fixed size list type
-      const vectorColumnOptions = opt.vectorColumns[colName];
-      if (vectorColumnOptions !== undefined) {
-        const firstNonNullValue = values.find((v) => v !== null);
-        if (Array.isArray(firstNonNullValue)) {
-          type = newVectorType(
-            firstNonNullValue.length,
-            vectorColumnOptions.type,
-          );
-        } else {
-          throw new Error(
-            `Column ${colName} is expected to be a vector column but first non-null value is not an array.  Could not determine size of vector column`,
-          );
-        }
-      }
-    }
+		// By default (type === undefined) arrow will infer the type from the JS type
+		let type;
+		if (opt.schema !== undefined) {
+			// If there is a schema provided, then use that for the type instead
+			type = opt.schema?.fields.filter((f) => f.name === colName)[0]?.type;
+			if (DataType.isInt(type) && type.bitWidth === 64) {
+				// wrap in BigInt to avoid bug: https://github.com/apache/arrow/issues/40051
+				values = values.map((v) => {
+					if (v === null) {
+						return v;
+					}
+					if (typeof v === "bigint") {
+						return v;
+					}
+					if (typeof v === "number") {
+						return BigInt(v);
+					}
+					throw new Error(
+						`Expected BigInt or number for column ${colName}, got ${typeof v}`,
+					);
+				});
+			}
+		} else {
+			// Otherwise, check to see if this column is one of the vector columns
+			// defined by opt.vectorColumns and, if so, use the fixed size list type
+			const vectorColumnOptions = opt.vectorColumns[colName];
+			if (vectorColumnOptions !== undefined) {
+				const firstNonNullValue = values.find((v) => v !== null);
+				if (Array.isArray(firstNonNullValue)) {
+					type = newVectorType(
+						firstNonNullValue.length,
+						vectorColumnOptions.type,
+					);
+				} else {
+					throw new Error(
+						`Column ${colName} is expected to be a vector column but first non-null value is not an array.  Could not determine size of vector column`,
+					);
+				}
+			}
+		}
 
-    try {
-      // Convert an Array of JS values to an arrow vector
-      columns[colName] = makeVector(values, type, opt.dictionaryEncodeStrings);
-    } catch (error: unknown) {
-      // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-      throw Error(`Could not convert column "${colName}" to Arrow: ${error}`);
-    }
-  }
+		try {
+			// Convert an Array of JS values to an arrow vector
+			columns[colName] = makeVector(values, type, opt.dictionaryEncodeStrings);
+		} catch (error: unknown) {
+			// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+			throw Error(`Could not convert column "${colName}" to Arrow: ${error}`);
+		}
+	}
 
-  if (opt.schema != null) {
-    // `new ArrowTable(columns)` infers a schema which may sometimes have
-    // incorrect nullability (it assumes nullable=true always)
-    //
-    // `new ArrowTable(schema, columns)` will also fail because it will create a
-    // batch with an inferred schema and then complain that the batch schema
-    // does not match the provided schema.
-    //
-    // To work around this we first create a table with the wrong schema and
-    // then patch the schema of the batches so we can use
-    // `new ArrowTable(schema, batches)` which does not do any schema inference
-    const firstTable = new ArrowTable(columns);
-    const batchesFixed = firstTable.batches.map(
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      (batch) => new RecordBatch(opt.schema!, batch.data),
-    );
-    return new ArrowTable(opt.schema, batchesFixed);
-  } else {
-    return new ArrowTable(columns);
-  }
+	if (opt.schema != null) {
+		// `new ArrowTable(columns)` infers a schema which may sometimes have
+		// incorrect nullability (it assumes nullable=true always)
+		//
+		// `new ArrowTable(schema, columns)` will also fail because it will create a
+		// batch with an inferred schema and then complain that the batch schema
+		// does not match the provided schema.
+		//
+		// To work around this we first create a table with the wrong schema and
+		// then patch the schema of the batches so we can use
+		// `new ArrowTable(schema, batches)` which does not do any schema inference
+		const firstTable = new ArrowTable(columns);
+		const batchesFixed = firstTable.batches.map(
+			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+			(batch) => new RecordBatch(opt.schema!, batch.data),
+		);
+		return new ArrowTable(opt.schema, batchesFixed);
+	} else {
+		return new ArrowTable(columns);
+	}
 }
 
 /**
  * Create an empty Arrow table with the provided schema
  */
 export function makeEmptyTable(schema: Schema): ArrowTable {
-  return makeArrowTable([], { schema });
+	return makeArrowTable([], { schema });
 }
 
 /**
@@ -311,161 +311,161 @@ export function makeEmptyTable(schema: Schema): ArrowTable {
  */
 // @ts-expect-error (Vector<unknown> is not assignable to Vector<any>)
 function makeListVector(lists: unknown[][]): Vector<unknown> {
-  if (lists.length === 0 || lists[0].length === 0) {
-    throw Error("Cannot infer list vector from empty array or empty list");
-  }
-  const sampleList = lists[0];
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let inferredType: any;
-  try {
-    const sampleVector = makeVector(sampleList);
-    inferredType = sampleVector.type;
-  } catch (error: unknown) {
-    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-    throw Error(`Cannot infer list vector.  Cannot infer inner type: ${error}`);
-  }
+	if (lists.length === 0 || lists[0].length === 0) {
+		throw Error("Cannot infer list vector from empty array or empty list");
+	}
+	const sampleList = lists[0];
+	// biome-ignore lint/suspicious/noExplicitAny: skip
+	let inferredType: any;
+	try {
+		const sampleVector = makeVector(sampleList);
+		inferredType = sampleVector.type;
+	} catch (error: unknown) {
+		// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+		throw Error(`Cannot infer list vector.  Cannot infer inner type: ${error}`);
+	}
 
-  const listBuilder = makeBuilder({
-    type: new List(new Field("item", inferredType, true)),
-  });
-  for (const list of lists) {
-    listBuilder.append(list);
-  }
-  return listBuilder.finish().toVector();
+	const listBuilder = makeBuilder({
+		type: new List(new Field("item", inferredType, true)),
+	});
+	for (const list of lists) {
+		listBuilder.append(list);
+	}
+	return listBuilder.finish().toVector();
 }
 
 /** Helper function to convert an Array of JS values to an Arrow Vector */
 function makeVector(
-  values: unknown[],
-  type?: DataType,
-  stringAsDictionary?: boolean,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+	values: unknown[],
+	type?: DataType,
+	stringAsDictionary?: boolean,
+	// biome-ignore lint/suspicious/noExplicitAny: skip
 ): Vector<any> {
-  if (type !== undefined) {
-    // No need for inference, let Arrow create it
-    return vectorFromArray(values, type);
-  }
-  if (values.length === 0) {
-    throw Error(
-      "makeVector requires at least one value or the type must be specfied",
-    );
-  }
-  const sampleValue = values.find((val) => val !== null && val !== undefined);
-  if (sampleValue === undefined) {
-    throw Error(
-      "makeVector cannot infer the type if all values are null or undefined",
-    );
-  }
-  if (Array.isArray(sampleValue)) {
-    // Default Arrow inference doesn't handle list types
-    return makeListVector(values as unknown[][]);
-  } else if (Buffer.isBuffer(sampleValue)) {
-    // Default Arrow inference doesn't handle Buffer
-    return vectorFromArray(values, new Binary());
-  } else if (
-    !(stringAsDictionary ?? false) &&
-    (typeof sampleValue === "string" || sampleValue instanceof String)
-  ) {
-    // If the type is string then don't use Arrow's default inference unless dictionaries are requested
-    // because it will always use dictionary encoding for strings
-    return vectorFromArray(values, new Utf8());
-  } else {
-    // Convert a JS array of values to an arrow vector
-    return vectorFromArray(values);
-  }
+	if (type !== undefined) {
+		// No need for inference, let Arrow create it
+		return vectorFromArray(values, type);
+	}
+	if (values.length === 0) {
+		throw Error(
+			"makeVector requires at least one value or the type must be specfied",
+		);
+	}
+	const sampleValue = values.find((val) => val !== null && val !== undefined);
+	if (sampleValue === undefined) {
+		throw Error(
+			"makeVector cannot infer the type if all values are null or undefined",
+		);
+	}
+	if (Array.isArray(sampleValue)) {
+		// Default Arrow inference doesn't handle list types
+		return makeListVector(values as unknown[][]);
+	} else if (Buffer.isBuffer(sampleValue)) {
+		// Default Arrow inference doesn't handle Buffer
+		return vectorFromArray(values, new Binary());
+	} else if (
+		!(stringAsDictionary ?? false) &&
+		(typeof sampleValue === "string" || sampleValue instanceof String)
+	) {
+		// If the type is string then don't use Arrow's default inference unless dictionaries are requested
+		// because it will always use dictionary encoding for strings
+		return vectorFromArray(values, new Utf8());
+	} else {
+		// Convert a JS array of values to an arrow vector
+		return vectorFromArray(values);
+	}
 }
 
 /** Helper function to apply embeddings to an input table */
 async function applyEmbeddings<T>(
-  table: ArrowTable,
-  embeddings?: EmbeddingFunction<T>,
-  schema?: Schema,
+	table: ArrowTable,
+	embeddings?: EmbeddingFunction<T>,
+	schema?: Schema,
 ): Promise<ArrowTable> {
-  if (embeddings == null) {
-    return table;
-  }
+	if (embeddings == null) {
+		return table;
+	}
 
-  if (schema !== undefined && schema !== null) {
-    schema = sanitizeSchema(schema);
-  }
+	if (schema !== undefined && schema !== null) {
+		schema = sanitizeSchema(schema);
+	}
 
-  // Convert from ArrowTable to Record<String, Vector>
-  const colEntries = [...Array(table.numCols).keys()].map((_, idx) => {
-    const name = table.schema.fields[idx].name;
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const vec = table.getChildAt(idx)!;
-    return [name, vec];
-  });
-  const newColumns = Object.fromEntries(colEntries);
+	// Convert from ArrowTable to Record<String, Vector>
+	const colEntries = [...Array(table.numCols).keys()].map((_, idx) => {
+		const name = table.schema.fields[idx].name;
+		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+		const vec = table.getChildAt(idx)!;
+		return [name, vec];
+	});
+	const newColumns = Object.fromEntries(colEntries);
 
-  const sourceColumn = newColumns[embeddings.sourceColumn];
-  const destColumn = embeddings.destColumn ?? "vector";
-  const innerDestType = embeddings.embeddingDataType ?? new Float32();
-  if (sourceColumn === undefined) {
-    throw new Error(
-      `Cannot apply embedding function because the source column '${embeddings.sourceColumn}' was not present in the data`,
-    );
-  }
+	const sourceColumn = newColumns[embeddings.sourceColumn];
+	const destColumn = embeddings.destColumn ?? "vector";
+	const innerDestType = embeddings.embeddingDataType ?? new Float32();
+	if (sourceColumn === undefined) {
+		throw new Error(
+			`Cannot apply embedding function because the source column '${embeddings.sourceColumn}' was not present in the data`,
+		);
+	}
 
-  if (table.numRows === 0) {
-    if (Object.prototype.hasOwnProperty.call(newColumns, destColumn)) {
-      // We have an empty table and it already has the embedding column so no work needs to be done
-      // Note: we don't return an error like we did below because this is a common occurrence.  For example,
-      // if we call convertToTable with 0 records and a schema that includes the embedding
-      return table;
-    }
-    if (embeddings.embeddingDimension !== undefined) {
-      const destType = newVectorType(
-        embeddings.embeddingDimension,
-        innerDestType,
-      );
-      newColumns[destColumn] = makeVector([], destType);
-    } else if (schema != null) {
-      const destField = schema.fields.find((f) => f.name === destColumn);
-      if (destField != null) {
-        newColumns[destColumn] = makeVector([], destField.type);
-      } else {
-        throw new Error(
-          `Attempt to apply embeddings to an empty table failed because schema was missing embedding column '${destColumn}'`,
-        );
-      }
-    } else {
-      throw new Error(
-        "Attempt to apply embeddings to an empty table when the embeddings function does not specify `embeddingDimension`",
-      );
-    }
-  } else {
-    if (Object.prototype.hasOwnProperty.call(newColumns, destColumn)) {
-      throw new Error(
-        `Attempt to apply embeddings to table failed because column ${destColumn} already existed`,
-      );
-    }
-    if (table.batches.length > 1) {
-      throw new Error(
-        "Internal error: `makeArrowTable` unexpectedly created a table with more than one batch",
-      );
-    }
-    const values = sourceColumn.toArray();
-    const vectors = await embeddings.embed(values as T[]);
-    if (vectors.length !== values.length) {
-      throw new Error(
-        "Embedding function did not return an embedding for each input element",
-      );
-    }
-    const destType = newVectorType(vectors[0].length, innerDestType);
-    newColumns[destColumn] = makeVector(vectors, destType);
-  }
+	if (table.numRows === 0) {
+		if (Object.prototype.hasOwnProperty.call(newColumns, destColumn)) {
+			// We have an empty table and it already has the embedding column so no work needs to be done
+			// Note: we don't return an error like we did below because this is a common occurrence.  For example,
+			// if we call convertToTable with 0 records and a schema that includes the embedding
+			return table;
+		}
+		if (embeddings.embeddingDimension !== undefined) {
+			const destType = newVectorType(
+				embeddings.embeddingDimension,
+				innerDestType,
+			);
+			newColumns[destColumn] = makeVector([], destType);
+		} else if (schema != null) {
+			const destField = schema.fields.find((f) => f.name === destColumn);
+			if (destField != null) {
+				newColumns[destColumn] = makeVector([], destField.type);
+			} else {
+				throw new Error(
+					`Attempt to apply embeddings to an empty table failed because schema was missing embedding column '${destColumn}'`,
+				);
+			}
+		} else {
+			throw new Error(
+				"Attempt to apply embeddings to an empty table when the embeddings function does not specify `embeddingDimension`",
+			);
+		}
+	} else {
+		if (Object.prototype.hasOwnProperty.call(newColumns, destColumn)) {
+			throw new Error(
+				`Attempt to apply embeddings to table failed because column ${destColumn} already existed`,
+			);
+		}
+		if (table.batches.length > 1) {
+			throw new Error(
+				"Internal error: `makeArrowTable` unexpectedly created a table with more than one batch",
+			);
+		}
+		const values = sourceColumn.toArray();
+		const vectors = await embeddings.embed(values as T[]);
+		if (vectors.length !== values.length) {
+			throw new Error(
+				"Embedding function did not return an embedding for each input element",
+			);
+		}
+		const destType = newVectorType(vectors[0].length, innerDestType);
+		newColumns[destColumn] = makeVector(vectors, destType);
+	}
 
-  const newTable = new ArrowTable(newColumns);
-  if (schema != null) {
-    if (schema.fields.find((f) => f.name === destColumn) === undefined) {
-      throw new Error(
-        `When using embedding functions and specifying a schema the schema should include the embedding column but the column ${destColumn} was missing`,
-      );
-    }
-    return alignTable(newTable, schema);
-  }
-  return newTable;
+	const newTable = new ArrowTable(newColumns);
+	if (schema != null) {
+		if (schema.fields.find((f) => f.name === destColumn) === undefined) {
+			throw new Error(
+				`When using embedding functions and specifying a schema the schema should include the embedding column but the column ${destColumn} was missing`,
+			);
+		}
+		return alignTable(newTable, schema);
+	}
+	return newTable;
 }
 
 /**
@@ -487,23 +487,23 @@ async function applyEmbeddings<T>(
  * be placed at the end of the table, after all of the input columns.
  */
 export async function convertToTable<T>(
-  data: Array<Record<string, unknown>>,
-  embeddings?: EmbeddingFunction<T>,
-  makeTableOptions?: Partial<MakeArrowTableOptions>,
+	data: Array<Record<string, unknown>>,
+	embeddings?: EmbeddingFunction<T>,
+	makeTableOptions?: Partial<MakeArrowTableOptions>,
 ): Promise<ArrowTable> {
-  const table = makeArrowTable(data, makeTableOptions);
-  return await applyEmbeddings(table, embeddings, makeTableOptions?.schema);
+	const table = makeArrowTable(data, makeTableOptions);
+	return await applyEmbeddings(table, embeddings, makeTableOptions?.schema);
 }
 
 /** Creates the Arrow Type for a Vector column with dimension `dim` */
 function newVectorType<T extends Float>(
-  dim: number,
-  innerType: T,
+	dim: number,
+	innerType: T,
 ): FixedSizeList<T> {
-  // in Lance we always default to have the elements nullable, so we need to set it to true
-  // otherwise we often get schema mismatches because the stored data always has schema with nullable elements
-  const children = new Field<T>("item", innerType, true);
-  return new FixedSizeList(dim, children);
+	// in Lance we always default to have the elements nullable, so we need to set it to true
+	// otherwise we often get schema mismatches because the stored data always has schema with nullable elements
+	const children = new Field<T>("item", innerType, true);
+	return new FixedSizeList(dim, children);
 }
 
 /**
@@ -514,16 +514,16 @@ function newVectorType<T extends Float>(
  * `schema` is required if data is empty
  */
 export async function fromRecordsToBuffer<T>(
-  data: Array<Record<string, unknown>>,
-  embeddings?: EmbeddingFunction<T>,
-  schema?: Schema,
+	data: Array<Record<string, unknown>>,
+	embeddings?: EmbeddingFunction<T>,
+	schema?: Schema,
 ): Promise<Buffer> {
-  if (schema !== undefined && schema !== null) {
-    schema = sanitizeSchema(schema);
-  }
-  const table = await convertToTable(data, embeddings, { schema });
-  const writer = RecordBatchFileWriter.writeAll(table);
-  return Buffer.from(await writer.toUint8Array());
+	if (schema !== undefined && schema !== null) {
+		schema = sanitizeSchema(schema);
+	}
+	const table = await convertToTable(data, embeddings, { schema });
+	const writer = RecordBatchFileWriter.writeAll(table);
+	return Buffer.from(await writer.toUint8Array());
 }
 
 /**
@@ -534,16 +534,16 @@ export async function fromRecordsToBuffer<T>(
  * `schema` is required if data is empty
  */
 export async function fromRecordsToStreamBuffer<T>(
-  data: Array<Record<string, unknown>>,
-  embeddings?: EmbeddingFunction<T>,
-  schema?: Schema,
+	data: Array<Record<string, unknown>>,
+	embeddings?: EmbeddingFunction<T>,
+	schema?: Schema,
 ): Promise<Buffer> {
-  if (schema !== undefined && schema !== null) {
-    schema = sanitizeSchema(schema);
-  }
-  const table = await convertToTable(data, embeddings, { schema });
-  const writer = RecordBatchStreamWriter.writeAll(table);
-  return Buffer.from(await writer.toUint8Array());
+	if (schema !== undefined && schema !== null) {
+		schema = sanitizeSchema(schema);
+	}
+	const table = await convertToTable(data, embeddings, { schema });
+	const writer = RecordBatchStreamWriter.writeAll(table);
+	return Buffer.from(await writer.toUint8Array());
 }
 
 /**
@@ -555,16 +555,16 @@ export async function fromRecordsToStreamBuffer<T>(
  * `schema` is required if the table is empty
  */
 export async function fromTableToBuffer<T>(
-  table: ArrowTable,
-  embeddings?: EmbeddingFunction<T>,
-  schema?: Schema,
+	table: ArrowTable,
+	embeddings?: EmbeddingFunction<T>,
+	schema?: Schema,
 ): Promise<Buffer> {
-  if (schema !== undefined && schema !== null) {
-    schema = sanitizeSchema(schema);
-  }
-  const tableWithEmbeddings = await applyEmbeddings(table, embeddings, schema);
-  const writer = RecordBatchFileWriter.writeAll(tableWithEmbeddings);
-  return Buffer.from(await writer.toUint8Array());
+	if (schema !== undefined && schema !== null) {
+		schema = sanitizeSchema(schema);
+	}
+	const tableWithEmbeddings = await applyEmbeddings(table, embeddings, schema);
+	const writer = RecordBatchFileWriter.writeAll(tableWithEmbeddings);
+	return Buffer.from(await writer.toUint8Array());
 }
 
 /**
@@ -576,19 +576,19 @@ export async function fromTableToBuffer<T>(
  * `schema` is required if the table is empty
  */
 export async function fromDataToBuffer<T>(
-  data: Data,
-  embeddings?: EmbeddingFunction<T>,
-  schema?: Schema,
+	data: Data,
+	embeddings?: EmbeddingFunction<T>,
+	schema?: Schema,
 ): Promise<Buffer> {
-  if (schema !== undefined && schema !== null) {
-    schema = sanitizeSchema(schema);
-  }
-  if (data instanceof ArrowTable) {
-    return fromTableToBuffer(data, embeddings, schema);
-  } else {
-    const table = await convertToTable(data);
-    return fromTableToBuffer(table, embeddings, schema);
-  }
+	if (schema !== undefined && schema !== null) {
+		schema = sanitizeSchema(schema);
+	}
+	if (data instanceof ArrowTable) {
+		return fromTableToBuffer(data, embeddings, schema);
+	} else {
+		const table = await convertToTable(data);
+		return fromTableToBuffer(table, embeddings, schema);
+	}
 }
 
 /**
@@ -600,91 +600,91 @@ export async function fromDataToBuffer<T>(
  * `schema` is required if the table is empty
  */
 export async function fromTableToStreamBuffer<T>(
-  table: ArrowTable,
-  embeddings?: EmbeddingFunction<T>,
-  schema?: Schema,
+	table: ArrowTable,
+	embeddings?: EmbeddingFunction<T>,
+	schema?: Schema,
 ): Promise<Buffer> {
-  const tableWithEmbeddings = await applyEmbeddings(table, embeddings, schema);
-  const writer = RecordBatchStreamWriter.writeAll(tableWithEmbeddings);
-  return Buffer.from(await writer.toUint8Array());
+	const tableWithEmbeddings = await applyEmbeddings(table, embeddings, schema);
+	const writer = RecordBatchStreamWriter.writeAll(tableWithEmbeddings);
+	return Buffer.from(await writer.toUint8Array());
 }
 
 /**
  * Reorder the columns in `batch` so that they agree with the field order in `schema`
  */
 function alignBatch(batch: RecordBatch, schema: Schema): RecordBatch {
-  const alignedChildren = [];
-  for (const field of schema.fields) {
-    const indexInBatch = batch.schema.fields?.findIndex(
-      (f) => f.name === field.name,
-    );
-    if (indexInBatch < 0) {
-      throw new Error(
-        `The column ${field.name} was not found in the Arrow Table`,
-      );
-    }
-    alignedChildren.push(batch.data.children[indexInBatch]);
-  }
-  const newData = makeData({
-    type: new Struct(schema.fields),
-    length: batch.numRows,
-    nullCount: batch.nullCount,
-    children: alignedChildren,
-  });
-  return new RecordBatch(schema, newData);
+	const alignedChildren = [];
+	for (const field of schema.fields) {
+		const indexInBatch = batch.schema.fields?.findIndex(
+			(f) => f.name === field.name,
+		);
+		if (indexInBatch < 0) {
+			throw new Error(
+				`The column ${field.name} was not found in the Arrow Table`,
+			);
+		}
+		alignedChildren.push(batch.data.children[indexInBatch]);
+	}
+	const newData = makeData({
+		type: new Struct(schema.fields),
+		length: batch.numRows,
+		nullCount: batch.nullCount,
+		children: alignedChildren,
+	});
+	return new RecordBatch(schema, newData);
 }
 
 /**
  * Reorder the columns in `table` so that they agree with the field order in `schema`
  */
 function alignTable(table: ArrowTable, schema: Schema): ArrowTable {
-  const alignedBatches = table.batches.map((batch) =>
-    alignBatch(batch, schema),
-  );
-  return new ArrowTable(schema, alignedBatches);
+	const alignedBatches = table.batches.map((batch) =>
+		alignBatch(batch, schema),
+	);
+	return new ArrowTable(schema, alignedBatches);
 }
 
 /**
  * Create an empty table with the given schema
  */
 export function createEmptyTable(schema: Schema): ArrowTable {
-  return new ArrowTable(sanitizeSchema(schema));
+	return new ArrowTable(sanitizeSchema(schema));
 }
 
 function validateSchemaEmbeddings(
-  schema: Schema,
-  data: Array<Record<string, unknown>>,
-  embeddings: EmbeddingFunction<unknown> | undefined,
+	schema: Schema,
+	data: Array<Record<string, unknown>>,
+	embeddings: EmbeddingFunction<unknown> | undefined,
 ) {
-  const fields = [];
-  const missingEmbeddingFields = [];
+	const fields = [];
+	const missingEmbeddingFields = [];
 
-  // First we check if the field is a `FixedSizeList`
-  // Then we check if the data contains the field
-  // if it does not, we add it to the list of missing embedding fields
-  // Finally, we check if those missing embedding fields are `this._embeddings`
-  // if they are not, we throw an error
-  for (const field of schema.fields) {
-    if (field.type instanceof FixedSizeList) {
-      if (data.length !== 0 && data?.[0]?.[field.name] === undefined) {
-        missingEmbeddingFields.push(field);
-      } else {
-        fields.push(field);
-      }
-    } else {
-      fields.push(field);
-    }
-  }
+	// First we check if the field is a `FixedSizeList`
+	// Then we check if the data contains the field
+	// if it does not, we add it to the list of missing embedding fields
+	// Finally, we check if those missing embedding fields are `this._embeddings`
+	// if they are not, we throw an error
+	for (const field of schema.fields) {
+		if (field.type instanceof FixedSizeList) {
+			if (data.length !== 0 && data?.[0]?.[field.name] === undefined) {
+				missingEmbeddingFields.push(field);
+			} else {
+				fields.push(field);
+			}
+		} else {
+			fields.push(field);
+		}
+	}
 
-  if (missingEmbeddingFields.length > 0 && embeddings === undefined) {
-    console.log({ missingEmbeddingFields, embeddings });
+	if (missingEmbeddingFields.length > 0 && embeddings === undefined) {
+		console.log({ missingEmbeddingFields, embeddings });
 
-    throw new Error(
-      `Table has embeddings: "${missingEmbeddingFields
-        .map((f) => f.name)
-        .join(",")}", but no embedding function was provided`,
-    );
-  }
+		throw new Error(
+			`Table has embeddings: "${missingEmbeddingFields
+				.map((f) => f.name)
+				.join(",")}", but no embedding function was provided`,
+		);
+	}
 
-  return new Schema(fields);
+	return new Schema(fields);
 }
