+++ nodejs/lancedb/connection.ts
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { Table as ArrowTable, Schema } from "apache-arrow";
 import { fromTableToBuffer, makeArrowTable, makeEmptyTable } from "./arrow";
 import { ConnectionOptions, Connection as LanceDbConnection } from "./native";
 import { Table } from "./table";
-import { Table as ArrowTable, Schema } from "apache-arrow";
 
 /**
  * Connect to a LanceDB instance at the given URI.
@@ -30,78 +30,78 @@ import { Table as ArrowTable, Schema } from "apache-arrow";
  * @see {@link ConnectionOptions} for more details on the URI format.
  */
 export async function connect(
-  uri: string,
-  opts?: Partial<ConnectionOptions>,
+	uri: string,
+	opts?: Partial<ConnectionOptions>,
 ): Promise<Connection> {
-  opts = opts ?? {};
-  opts.storageOptions = cleanseStorageOptions(opts.storageOptions);
-  const nativeConn = await LanceDbConnection.new(uri, opts);
-  return new Connection(nativeConn);
+	opts = opts ?? {};
+	opts.storageOptions = cleanseStorageOptions(opts.storageOptions);
+	const nativeConn = await LanceDbConnection.new(uri, opts);
+	return new Connection(nativeConn);
 }
 
 export interface CreateTableOptions {
-  /**
-   * The mode to use when creating the table.
-   *
-   * If this is set to "create" and the table already exists then either
-   * an error will be thrown or, if existOk is true, then nothing will
-   * happen.  Any provided data will be ignored.
-   *
-   * If this is set to "overwrite" then any existing table will be replaced.
-   */
-  mode: "create" | "overwrite";
-  /**
-   * If this is true and the table already exists and the mode is "create"
-   * then no error will be raised.
-   */
-  existOk: boolean;
+	/**
+	 * The mode to use when creating the table.
+	 *
+	 * If this is set to "create" and the table already exists then either
+	 * an error will be thrown or, if existOk is true, then nothing will
+	 * happen.  Any provided data will be ignored.
+	 *
+	 * If this is set to "overwrite" then any existing table will be replaced.
+	 */
+	mode: "create" | "overwrite";
+	/**
+	 * If this is true and the table already exists and the mode is "create"
+	 * then no error will be raised.
+	 */
+	existOk: boolean;
 
-  /**
-   * Configuration for object storage.
-   *
-   * Options already set on the connection will be inherited by the table,
-   * but can be overridden here.
-   *
-   * The available options are described at https://lancedb.github.io/lancedb/guides/storage/
-   */
-  storageOptions?: Record<string, string>;
+	/**
+	 * Configuration for object storage.
+	 *
+	 * Options already set on the connection will be inherited by the table,
+	 * but can be overridden here.
+	 *
+	 * The available options are described at https://lancedb.github.io/lancedb/guides/storage/
+	 */
+	storageOptions?: Record<string, string>;
 }
 
 export interface OpenTableOptions {
-  /**
-   * Configuration for object storage.
-   *
-   * Options already set on the connection will be inherited by the table,
-   * but can be overridden here.
-   *
-   * The available options are described at https://lancedb.github.io/lancedb/guides/storage/
-   */
-  storageOptions?: Record<string, string>;
-  /**
-   * Set the size of the index cache, specified as a number of entries
-   *
-   * The exact meaning of an "entry" will depend on the type of index:
-   * - IVF: there is one entry for each IVF partition
-   * - BTREE: there is one entry for the entire index
-   *
-   * This cache applies to the entire opened table, across all indices.
-   * Setting this value higher will increase performance on larger datasets
-   * at the expense of more RAM
-   */
-  indexCacheSize?: number;
+	/**
+	 * Configuration for object storage.
+	 *
+	 * Options already set on the connection will be inherited by the table,
+	 * but can be overridden here.
+	 *
+	 * The available options are described at https://lancedb.github.io/lancedb/guides/storage/
+	 */
+	storageOptions?: Record<string, string>;
+	/**
+	 * Set the size of the index cache, specified as a number of entries
+	 *
+	 * The exact meaning of an "entry" will depend on the type of index:
+	 * - IVF: there is one entry for each IVF partition
+	 * - BTREE: there is one entry for the entire index
+	 *
+	 * This cache applies to the entire opened table, across all indices.
+	 * Setting this value higher will increase performance on larger datasets
+	 * at the expense of more RAM
+	 */
+	indexCacheSize?: number;
 }
 
 export interface TableNamesOptions {
-  /**
-   * If present, only return names that come lexicographically after the
-   * supplied value.
-   *
-   * This can be combined with limit to implement pagination by setting this to
-   * the last table name from the previous page.
-   */
-  startAfter?: string;
-  /** An optional limit to the number of results to return. */
-  limit?: number;
+	/**
+	 * If present, only return names that come lexicographically after the
+	 * supplied value.
+	 *
+	 * This can be combined with limit to implement pagination by setting this to
+	 * the last table name from the previous page.
+	 */
+	startAfter?: string;
+	/** An optional limit to the number of results to return. */
+	limit?: number;
 }
 
 /**
@@ -123,148 +123,148 @@ export interface TableNamesOptions {
  * the underlying connection has been closed.
  */
 export class Connection {
-  readonly inner: LanceDbConnection;
+	readonly inner: LanceDbConnection;
 
-  constructor(inner: LanceDbConnection) {
-    this.inner = inner;
-  }
+	constructor(inner: LanceDbConnection) {
+		this.inner = inner;
+	}
 
-  /** Return true if the connection has not been closed */
-  isOpen(): boolean {
-    return this.inner.isOpen();
-  }
+	/** Return true if the connection has not been closed */
+	isOpen(): boolean {
+		return this.inner.isOpen();
+	}
 
-  /**
-   * Close the connection, releasing any underlying resources.
-   *
-   * It is safe to call this method multiple times.
-   *
-   * Any attempt to use the connection after it is closed will result in an error.
-   */
-  close(): void {
-    this.inner.close();
-  }
+	/**
+	 * Close the connection, releasing any underlying resources.
+	 *
+	 * It is safe to call this method multiple times.
+	 *
+	 * Any attempt to use the connection after it is closed will result in an error.
+	 */
+	close(): void {
+		this.inner.close();
+	}
 
-  /** Return a brief description of the connection */
-  display(): string {
-    return this.inner.display();
-  }
+	/** Return a brief description of the connection */
+	display(): string {
+		return this.inner.display();
+	}
 
-  /**
-   * List all the table names in this database.
-   *
-   * Tables will be returned in lexicographical order.
-   * @param {Partial<TableNamesOptions>} options - options to control the
-   * paging / start point
-   */
-  async tableNames(options?: Partial<TableNamesOptions>): Promise<string[]> {
-    return this.inner.tableNames(options?.startAfter, options?.limit);
-  }
+	/**
+	 * List all the table names in this database.
+	 *
+	 * Tables will be returned in lexicographical order.
+	 * @param {Partial<TableNamesOptions>} options - options to control the
+	 * paging / start point
+	 */
+	async tableNames(options?: Partial<TableNamesOptions>): Promise<string[]> {
+		return this.inner.tableNames(options?.startAfter, options?.limit);
+	}
 
-  /**
-   * Open a table in the database.
-   * @param {string} name - The name of the table
-   */
-  async openTable(
-    name: string,
-    options?: Partial<OpenTableOptions>,
-  ): Promise<Table> {
-    const innerTable = await this.inner.openTable(
-      name,
-      cleanseStorageOptions(options?.storageOptions),
-      options?.indexCacheSize,
-    );
-    return new Table(innerTable);
-  }
+	/**
+	 * Open a table in the database.
+	 * @param {string} name - The name of the table
+	 */
+	async openTable(
+		name: string,
+		options?: Partial<OpenTableOptions>,
+	): Promise<Table> {
+		const innerTable = await this.inner.openTable(
+			name,
+			cleanseStorageOptions(options?.storageOptions),
+			options?.indexCacheSize,
+		);
+		return new Table(innerTable);
+	}
 
-  /**
-   * Creates a new Table and initialize it with new data.
-   * @param {string} name - The name of the table.
-   * @param {Record<string, unknown>[] | ArrowTable} data - Non-empty Array of Records
-   * to be inserted into the table
-   */
-  async createTable(
-    name: string,
-    data: Record<string, unknown>[] | ArrowTable,
-    options?: Partial<CreateTableOptions>,
-  ): Promise<Table> {
-    let mode: string = options?.mode ?? "create";
-    const existOk = options?.existOk ?? false;
+	/**
+	 * Creates a new Table and initialize it with new data.
+	 * @param {string} name - The name of the table.
+	 * @param {Record<string, unknown>[] | ArrowTable} data - Non-empty Array of Records
+	 * to be inserted into the table
+	 */
+	async createTable(
+		name: string,
+		data: Record<string, unknown>[] | ArrowTable,
+		options?: Partial<CreateTableOptions>,
+	): Promise<Table> {
+		let mode: string = options?.mode ?? "create";
+		const existOk = options?.existOk ?? false;
 
-    if (mode === "create" && existOk) {
-      mode = "exist_ok";
-    }
+		if (mode === "create" && existOk) {
+			mode = "exist_ok";
+		}
 
-    let table: ArrowTable;
-    if (data instanceof ArrowTable) {
-      table = data;
-    } else {
-      table = makeArrowTable(data);
-    }
-    const buf = await fromTableToBuffer(table);
-    const innerTable = await this.inner.createTable(
-      name,
-      buf,
-      mode,
-      cleanseStorageOptions(options?.storageOptions),
-    );
-    return new Table(innerTable);
-  }
+		let table: ArrowTable;
+		if (data instanceof ArrowTable) {
+			table = data;
+		} else {
+			table = makeArrowTable(data);
+		}
+		const buf = await fromTableToBuffer(table);
+		const innerTable = await this.inner.createTable(
+			name,
+			buf,
+			mode,
+			cleanseStorageOptions(options?.storageOptions),
+		);
+		return new Table(innerTable);
+	}
 
-  /**
-   * Creates a new empty Table
-   * @param {string} name - The name of the table.
-   * @param {Schema} schema - The schema of the table
-   */
-  async createEmptyTable(
-    name: string,
-    schema: Schema,
-    options?: Partial<CreateTableOptions>,
-  ): Promise<Table> {
-    let mode: string = options?.mode ?? "create";
-    const existOk = options?.existOk ?? false;
+	/**
+	 * Creates a new empty Table
+	 * @param {string} name - The name of the table.
+	 * @param {Schema} schema - The schema of the table
+	 */
+	async createEmptyTable(
+		name: string,
+		schema: Schema,
+		options?: Partial<CreateTableOptions>,
+	): Promise<Table> {
+		let mode: string = options?.mode ?? "create";
+		const existOk = options?.existOk ?? false;
 
-    if (mode === "create" && existOk) {
-      mode = "exist_ok";
-    }
+		if (mode === "create" && existOk) {
+			mode = "exist_ok";
+		}
 
-    const table = makeEmptyTable(schema);
-    const buf = await fromTableToBuffer(table);
-    const innerTable = await this.inner.createEmptyTable(
-      name,
-      buf,
-      mode,
-      cleanseStorageOptions(options?.storageOptions),
-    );
-    return new Table(innerTable);
-  }
+		const table = makeEmptyTable(schema);
+		const buf = await fromTableToBuffer(table);
+		const innerTable = await this.inner.createEmptyTable(
+			name,
+			buf,
+			mode,
+			cleanseStorageOptions(options?.storageOptions),
+		);
+		return new Table(innerTable);
+	}
 
-  /**
-   * Drop an existing table.
-   * @param {string} name The name of the table to drop.
-   */
-  async dropTable(name: string): Promise<void> {
-    return this.inner.dropTable(name);
-  }
+	/**
+	 * Drop an existing table.
+	 * @param {string} name The name of the table to drop.
+	 */
+	async dropTable(name: string): Promise<void> {
+		return this.inner.dropTable(name);
+	}
 }
 
 /**
  * Takes storage options and makes all the keys snake case.
  */
 function cleanseStorageOptions(
-  options?: Record<string, string>,
+	options?: Record<string, string>,
 ): Record<string, string> | undefined {
-  if (options === undefined) {
-    return undefined;
-  }
-  const result: Record<string, string> = {};
-  for (const [key, value] of Object.entries(options)) {
-    if (value !== undefined) {
-      const newKey = camelToSnakeCase(key);
-      result[newKey] = value;
-    }
-  }
-  return result;
+	if (options === undefined) {
+		return undefined;
+	}
+	const result: Record<string, string> = {};
+	for (const [key, value] of Object.entries(options)) {
+		if (value !== undefined) {
+			const newKey = camelToSnakeCase(key);
+			result[newKey] = value;
+		}
+	}
+	return result;
 }
 
 /**
@@ -272,18 +272,18 @@ function cleanseStorageOptions(
  * returned unchanged.
  */
 function camelToSnakeCase(camel: string): string {
-  if (camel.includes("_")) {
-    // Assume if there is at least one underscore, it is already snake case
-    return camel;
-  }
-  if (camel.toLocaleUpperCase() === camel) {
-    // Assume if the string is all uppercase, it is already snake case
-    return camel;
-  }
+	if (camel.includes("_")) {
+		// Assume if there is at least one underscore, it is already snake case
+		return camel;
+	}
+	if (camel.toLocaleUpperCase() === camel) {
+		// Assume if the string is all uppercase, it is already snake case
+		return camel;
+	}
 
-  let result = camel.replace(/[A-Z]/g, (letter) => `_${letter.toLowerCase()}`);
-  if (result.startsWith("_")) {
-    result = result.slice(1);
-  }
-  return result;
+	let result = camel.replace(/[A-Z]/g, (letter) => `_${letter.toLowerCase()}`);
+	if (result.startsWith("_")) {
+		result = result.slice(1);
+	}
+	return result;
 }
