+++ nodejs/lancedb/table.ts
@@ -13,41 +13,41 @@
 // limitations under the License.
 
 import { Schema, tableFromIPC } from "apache-arrow";
+import { Data, fromDataToBuffer } from "./arrow";
+import { IndexOptions } from "./indices";
 import {
-  AddColumnsSql,
-  ColumnAlteration,
-  IndexConfig,
-  Table as _NativeTable,
+	AddColumnsSql,
+	ColumnAlteration,
+	IndexConfig,
+	Table as _NativeTable,
 } from "./native";
 import { Query, VectorQuery } from "./query";
-import { IndexOptions } from "./indices";
-import { Data, fromDataToBuffer } from "./arrow";
 
 export { IndexConfig } from "./native";
 /**
  * Options for adding data to a table.
  */
 export interface AddDataOptions {
-  /**
-   * If "append" (the default) then the new data will be added to the table
-   *
-   * If "overwrite" then the new data will replace the existing data in the table.
-   */
-  mode: "append" | "overwrite";
+	/**
+	 * If "append" (the default) then the new data will be added to the table
+	 *
+	 * If "overwrite" then the new data will replace the existing data in the table.
+	 */
+	mode: "append" | "overwrite";
 }
 
 export interface UpdateOptions {
-  /**
-   * A filter that limits the scope of the update.
-   *
-   * This should be an SQL filter expression.
-   *
-   * Only rows that satisfy the expression will be updated.
-   *
-   * For example, this could be 'my_col == 0' to replace all instances
-   * of 0 in a column with some other default value.
-   */
-  where: string;
+	/**
+	 * A filter that limits the scope of the update.
+	 *
+	 * This should be an SQL filter expression.
+	 *
+	 * Only rows that satisfy the expression will be updated.
+	 *
+	 * For example, this could be 'my_col == 0' to replace all instances
+	 * of 0 in a column with some other default value.
+	 */
+	where: string;
 }
 
 /**
@@ -63,297 +63,297 @@ export interface UpdateOptions {
  * collected.
  */
 export class Table {
-  private readonly inner: _NativeTable;
+	private readonly inner: _NativeTable;
 
-  /** Construct a Table. Internal use only. */
-  constructor(inner: _NativeTable) {
-    this.inner = inner;
-  }
+	/** Construct a Table. Internal use only. */
+	constructor(inner: _NativeTable) {
+		this.inner = inner;
+	}
 
-  /** Return true if the table has not been closed */
-  isOpen(): boolean {
-    return this.inner.isOpen();
-  }
+	/** Return true if the table has not been closed */
+	isOpen(): boolean {
+		return this.inner.isOpen();
+	}
 
-  /**
-   * Close the table, releasing any underlying resources.
-   *
-   * It is safe to call this method multiple times.
-   *
-   * Any attempt to use the table after it is closed will result in an error.
-   */
-  close(): void {
-    this.inner.close();
-  }
+	/**
+	 * Close the table, releasing any underlying resources.
+	 *
+	 * It is safe to call this method multiple times.
+	 *
+	 * Any attempt to use the table after it is closed will result in an error.
+	 */
+	close(): void {
+		this.inner.close();
+	}
 
-  /** Return a brief description of the table */
-  display(): string {
-    return this.inner.display();
-  }
+	/** Return a brief description of the table */
+	display(): string {
+		return this.inner.display();
+	}
 
-  /** Get the schema of the table. */
-  async schema(): Promise<Schema> {
-    const schemaBuf = await this.inner.schema();
-    const tbl = tableFromIPC(schemaBuf);
-    return tbl.schema;
-  }
+	/** Get the schema of the table. */
+	async schema(): Promise<Schema> {
+		const schemaBuf = await this.inner.schema();
+		const tbl = tableFromIPC(schemaBuf);
+		return tbl.schema;
+	}
 
-  /**
-   * Insert records into this Table.
-   * @param {Data} data Records to be inserted into the Table
-   */
-  async add(data: Data, options?: Partial<AddDataOptions>): Promise<void> {
-    const mode = options?.mode ?? "append";
+	/**
+	 * Insert records into this Table.
+	 * @param {Data} data Records to be inserted into the Table
+	 */
+	async add(data: Data, options?: Partial<AddDataOptions>): Promise<void> {
+		const mode = options?.mode ?? "append";
 
-    const buffer = await fromDataToBuffer(data);
-    await this.inner.add(buffer, mode);
-  }
+		const buffer = await fromDataToBuffer(data);
+		await this.inner.add(buffer, mode);
+	}
 
-  /**
-   * Update existing records in the Table
-   *
-   * An update operation can be used to adjust existing values.  Use the
-   * returned builder to specify which columns to update.  The new value
-   * can be a literal value (e.g. replacing nulls with some default value)
-   * or an expression applied to the old value (e.g. incrementing a value)
-   *
-   * An optional condition can be specified (e.g. "only update if the old
-   * value is 0")
-   *
-   * Note: if your condition is something like "some_id_column == 7" and
-   * you are updating many rows (with different ids) then you will get
-   * better performance with a single [`merge_insert`] call instead of
-   * repeatedly calilng this method.
-   * @param {Map<string, string> | Record<string, string>} updates - the
-   * columns to update
-   *
-   * Keys in the map should specify the name of the column to update.
-   * Values in the map provide the new value of the column.  These can
-   * be SQL literal strings (e.g. "7" or "'foo'") or they can be expressions
-   * based on the row being updated (e.g. "my_col + 1")
-   * @param {Partial<UpdateOptions>} options - additional options to control
-   * the update behavior
-   */
-  async update(
-    updates: Map<string, string> | Record<string, string>,
-    options?: Partial<UpdateOptions>,
-  ) {
-    const onlyIf = options?.where;
-    let columns: [string, string][];
-    if (updates instanceof Map) {
-      columns = Array.from(updates.entries());
-    } else {
-      columns = Object.entries(updates);
-    }
-    await this.inner.update(onlyIf, columns);
-  }
+	/**
+	 * Update existing records in the Table
+	 *
+	 * An update operation can be used to adjust existing values.  Use the
+	 * returned builder to specify which columns to update.  The new value
+	 * can be a literal value (e.g. replacing nulls with some default value)
+	 * or an expression applied to the old value (e.g. incrementing a value)
+	 *
+	 * An optional condition can be specified (e.g. "only update if the old
+	 * value is 0")
+	 *
+	 * Note: if your condition is something like "some_id_column == 7" and
+	 * you are updating many rows (with different ids) then you will get
+	 * better performance with a single [`merge_insert`] call instead of
+	 * repeatedly calilng this method.
+	 * @param {Map<string, string> | Record<string, string>} updates - the
+	 * columns to update
+	 *
+	 * Keys in the map should specify the name of the column to update.
+	 * Values in the map provide the new value of the column.  These can
+	 * be SQL literal strings (e.g. "7" or "'foo'") or they can be expressions
+	 * based on the row being updated (e.g. "my_col + 1")
+	 * @param {Partial<UpdateOptions>} options - additional options to control
+	 * the update behavior
+	 */
+	async update(
+		updates: Map<string, string> | Record<string, string>,
+		options?: Partial<UpdateOptions>,
+	) {
+		const onlyIf = options?.where;
+		let columns: [string, string][];
+		if (updates instanceof Map) {
+			columns = Array.from(updates.entries());
+		} else {
+			columns = Object.entries(updates);
+		}
+		await this.inner.update(onlyIf, columns);
+	}
 
-  /** Count the total number of rows in the dataset. */
-  async countRows(filter?: string): Promise<number> {
-    return await this.inner.countRows(filter);
-  }
+	/** Count the total number of rows in the dataset. */
+	async countRows(filter?: string): Promise<number> {
+		return await this.inner.countRows(filter);
+	}
 
-  /** Delete the rows that satisfy the predicate. */
-  async delete(predicate: string): Promise<void> {
-    await this.inner.delete(predicate);
-  }
+	/** Delete the rows that satisfy the predicate. */
+	async delete(predicate: string): Promise<void> {
+		await this.inner.delete(predicate);
+	}
 
-  /**
-   * Create an index to speed up queries.
-   *
-   * Indices can be created on vector columns or scalar columns.
-   * Indices on vector columns will speed up vector searches.
-   * Indices on scalar columns will speed up filtering (in both
-   * vector and non-vector searches)
-   * @example
-   * // If the column has a vector (fixed size list) data type then
-   * // an IvfPq vector index will be created.
-   * const table = await conn.openTable("my_table");
-   * await table.createIndex("vector");
-   * @example
-   * // For advanced control over vector index creation you can specify
-   * // the index type and options.
-   * const table = await conn.openTable("my_table");
-   * await table.createIndex("vector", {
-   *   config: lancedb.Index.ivfPq({
-   *     numPartitions: 128,
-   *     numSubVectors: 16,
-   *   }),
-   * });
-   * @example
-   * // Or create a Scalar index
-   * await table.createIndex("my_float_col");
-   */
-  async createIndex(column: string, options?: Partial<IndexOptions>) {
-    // Bit of a hack to get around the fact that TS has no package-scope.
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const nativeIndex = (options?.config as any)?.inner;
-    await this.inner.createIndex(nativeIndex, column, options?.replace);
-  }
+	/**
+	 * Create an index to speed up queries.
+	 *
+	 * Indices can be created on vector columns or scalar columns.
+	 * Indices on vector columns will speed up vector searches.
+	 * Indices on scalar columns will speed up filtering (in both
+	 * vector and non-vector searches)
+	 * @example
+	 * // If the column has a vector (fixed size list) data type then
+	 * // an IvfPq vector index will be created.
+	 * const table = await conn.openTable("my_table");
+	 * await table.createIndex("vector");
+	 * @example
+	 * // For advanced control over vector index creation you can specify
+	 * // the index type and options.
+	 * const table = await conn.openTable("my_table");
+	 * await table.createIndex("vector", {
+	 *   config: lancedb.Index.ivfPq({
+	 *     numPartitions: 128,
+	 *     numSubVectors: 16,
+	 *   }),
+	 * });
+	 * @example
+	 * // Or create a Scalar index
+	 * await table.createIndex("my_float_col");
+	 */
+	async createIndex(column: string, options?: Partial<IndexOptions>) {
+		// Bit of a hack to get around the fact that TS has no package-scope.
+		// biome-ignore lint/suspicious/noExplicitAny: skip
+		const nativeIndex = (options?.config as any)?.inner;
+		await this.inner.createIndex(nativeIndex, column, options?.replace);
+	}
 
-  /**
-   * Create a {@link Query} Builder.
-   *
-   * Queries allow you to search your existing data.  By default the query will
-   * return all the data in the table in no particular order.  The builder
-   * returned by this method can be used to control the query using filtering,
-   * vector similarity, sorting, and more.
-   *
-   * Note: By default, all columns are returned.  For best performance, you should
-   * only fetch the columns you need.
-   *
-   * When appropriate, various indices and statistics based pruning will be used to
-   * accelerate the query.
-   * @example
-   * // SQL-style filtering
-   * //
-   * // This query will return up to 1000 rows whose value in the `id` column
-   * // is greater than 5. LanceDb supports a broad set of filtering functions.
-   * for await (const batch of table
-   *   .query()
-   *   .where("id > 1")
-   *   .select(["id"])
-   *   .limit(20)) {
-   *   console.log(batch);
-   * }
-   * @example
-   * // Vector Similarity Search
-   * //
-   * // This example will find the 10 rows whose value in the "vector" column are
-   * // closest to the query vector [1.0, 2.0, 3.0].  If an index has been created
-   * // on the "vector" column then this will perform an ANN search.
-   * //
-   * // The `refineFactor` and `nprobes` methods are used to control the recall /
-   * // latency tradeoff of the search.
-   * for await (const batch of table
-   *   .query()
-   *   .where("id > 1")
-   *   .select(["id"])
-   *   .limit(20)) {
-   *   console.log(batch);
-   * }
-   * @example
-   * // Scan the full dataset
-   * //
-   * // This query will return everything in the table in no particular order.
-   * for await (const batch of table.query()) {
-   *   console.log(batch);
-   * }
-   * @returns {Query} A builder that can be used to parameterize the query
-   */
-  query(): Query {
-    return new Query(this.inner);
-  }
+	/**
+	 * Create a {@link Query} Builder.
+	 *
+	 * Queries allow you to search your existing data.  By default the query will
+	 * return all the data in the table in no particular order.  The builder
+	 * returned by this method can be used to control the query using filtering,
+	 * vector similarity, sorting, and more.
+	 *
+	 * Note: By default, all columns are returned.  For best performance, you should
+	 * only fetch the columns you need.
+	 *
+	 * When appropriate, various indices and statistics based pruning will be used to
+	 * accelerate the query.
+	 * @example
+	 * // SQL-style filtering
+	 * //
+	 * // This query will return up to 1000 rows whose value in the `id` column
+	 * // is greater than 5. LanceDb supports a broad set of filtering functions.
+	 * for await (const batch of table
+	 *   .query()
+	 *   .where("id > 1")
+	 *   .select(["id"])
+	 *   .limit(20)) {
+	 *   console.log(batch);
+	 * }
+	 * @example
+	 * // Vector Similarity Search
+	 * //
+	 * // This example will find the 10 rows whose value in the "vector" column are
+	 * // closest to the query vector [1.0, 2.0, 3.0].  If an index has been created
+	 * // on the "vector" column then this will perform an ANN search.
+	 * //
+	 * // The `refineFactor` and `nprobes` methods are used to control the recall /
+	 * // latency tradeoff of the search.
+	 * for await (const batch of table
+	 *   .query()
+	 *   .where("id > 1")
+	 *   .select(["id"])
+	 *   .limit(20)) {
+	 *   console.log(batch);
+	 * }
+	 * @example
+	 * // Scan the full dataset
+	 * //
+	 * // This query will return everything in the table in no particular order.
+	 * for await (const batch of table.query()) {
+	 *   console.log(batch);
+	 * }
+	 * @returns {Query} A builder that can be used to parameterize the query
+	 */
+	query(): Query {
+		return new Query(this.inner);
+	}
 
-  /**
-   * Search the table with a given query vector.
-   *
-   * This is a convenience method for preparing a vector query and
-   * is the same thing as calling `nearestTo` on the builder returned
-   * by `query`.  @see {@link Query#nearestTo} for more details.
-   */
-  vectorSearch(vector: unknown): VectorQuery {
-    return this.query().nearestTo(vector);
-  }
+	/**
+	 * Search the table with a given query vector.
+	 *
+	 * This is a convenience method for preparing a vector query and
+	 * is the same thing as calling `nearestTo` on the builder returned
+	 * by `query`.  @see {@link Query#nearestTo} for more details.
+	 */
+	vectorSearch(vector: unknown): VectorQuery {
+		return this.query().nearestTo(vector);
+	}
 
-  // TODO: Support BatchUDF
-  /**
-   * Add new columns with defined values.
-   * @param {AddColumnsSql[]} newColumnTransforms pairs of column names and
-   * the SQL expression to use to calculate the value of the new column. These
-   * expressions will be evaluated for each row in the table, and can
-   * reference existing columns in the table.
-   */
-  async addColumns(newColumnTransforms: AddColumnsSql[]): Promise<void> {
-    await this.inner.addColumns(newColumnTransforms);
-  }
+	// TODO: Support BatchUDF
+	/**
+	 * Add new columns with defined values.
+	 * @param {AddColumnsSql[]} newColumnTransforms pairs of column names and
+	 * the SQL expression to use to calculate the value of the new column. These
+	 * expressions will be evaluated for each row in the table, and can
+	 * reference existing columns in the table.
+	 */
+	async addColumns(newColumnTransforms: AddColumnsSql[]): Promise<void> {
+		await this.inner.addColumns(newColumnTransforms);
+	}
 
-  /**
-   * Alter the name or nullability of columns.
-   * @param {ColumnAlteration[]} columnAlterations One or more alterations to
-   * apply to columns.
-   */
-  async alterColumns(columnAlterations: ColumnAlteration[]): Promise<void> {
-    await this.inner.alterColumns(columnAlterations);
-  }
+	/**
+	 * Alter the name or nullability of columns.
+	 * @param {ColumnAlteration[]} columnAlterations One or more alterations to
+	 * apply to columns.
+	 */
+	async alterColumns(columnAlterations: ColumnAlteration[]): Promise<void> {
+		await this.inner.alterColumns(columnAlterations);
+	}
 
-  /**
-   * Drop one or more columns from the dataset
-   *
-   * This is a metadata-only operation and does not remove the data from the
-   * underlying storage. In order to remove the data, you must subsequently
-   * call ``compact_files`` to rewrite the data without the removed columns and
-   * then call ``cleanup_files`` to remove the old files.
-   * @param {string[]} columnNames The names of the columns to drop. These can
-   * be nested column references (e.g. "a.b.c") or top-level column names
-   * (e.g. "a").
-   */
-  async dropColumns(columnNames: string[]): Promise<void> {
-    await this.inner.dropColumns(columnNames);
-  }
+	/**
+	 * Drop one or more columns from the dataset
+	 *
+	 * This is a metadata-only operation and does not remove the data from the
+	 * underlying storage. In order to remove the data, you must subsequently
+	 * call ``compact_files`` to rewrite the data without the removed columns and
+	 * then call ``cleanup_files`` to remove the old files.
+	 * @param {string[]} columnNames The names of the columns to drop. These can
+	 * be nested column references (e.g. "a.b.c") or top-level column names
+	 * (e.g. "a").
+	 */
+	async dropColumns(columnNames: string[]): Promise<void> {
+		await this.inner.dropColumns(columnNames);
+	}
 
-  /** Retrieve the version of the table */
-  async version(): Promise<number> {
-    return await this.inner.version();
-  }
+	/** Retrieve the version of the table */
+	async version(): Promise<number> {
+		return await this.inner.version();
+	}
 
-  /**
-   * Checks out a specific version of the table _This is an in-place operation._
-   *
-   * This allows viewing previous versions of the table. If you wish to
-   * keep writing to the dataset starting from an old version, then use
-   * the `restore` function.
-   *
-   * Calling this method will set the table into time-travel mode. If you
-   * wish to return to standard mode, call `checkoutLatest`.
-   * @param {number} version The version to checkout
-   * @example
-   * ```typescript
-   * import * as lancedb from "@lancedb/lancedb"
-   * const db = await lancedb.connect("./.lancedb");
-   * const table = await db.createTable("my_table", [
-   *   { vector: [1.1, 0.9], type: "vector" },
-   * ]);
-   *
-   * console.log(await table.version()); // 1
-   * console.log(table.display());
-   * await table.add([{ vector: [0.5, 0.2], type: "vector" }]);
-   * await table.checkout(1);
-   * console.log(await table.version()); // 2
-   * ```
-   */
-  async checkout(version: number): Promise<void> {
-    await this.inner.checkout(version);
-  }
+	/**
+	 * Checks out a specific version of the table _This is an in-place operation._
+	 *
+	 * This allows viewing previous versions of the table. If you wish to
+	 * keep writing to the dataset starting from an old version, then use
+	 * the `restore` function.
+	 *
+	 * Calling this method will set the table into time-travel mode. If you
+	 * wish to return to standard mode, call `checkoutLatest`.
+	 * @param {number} version The version to checkout
+	 * @example
+	 * ```typescript
+	 * import * as lancedb from "@lancedb/lancedb"
+	 * const db = await lancedb.connect("./.lancedb");
+	 * const table = await db.createTable("my_table", [
+	 *   { vector: [1.1, 0.9], type: "vector" },
+	 * ]);
+	 *
+	 * console.log(await table.version()); // 1
+	 * console.log(table.display());
+	 * await table.add([{ vector: [0.5, 0.2], type: "vector" }]);
+	 * await table.checkout(1);
+	 * console.log(await table.version()); // 2
+	 * ```
+	 */
+	async checkout(version: number): Promise<void> {
+		await this.inner.checkout(version);
+	}
 
-  /**
-   * Checkout the latest version of the table. _This is an in-place operation._
-   *
-   * The table will be set back into standard mode, and will track the latest
-   * version of the table.
-   */
-  async checkoutLatest(): Promise<void> {
-    await this.inner.checkoutLatest();
-  }
+	/**
+	 * Checkout the latest version of the table. _This is an in-place operation._
+	 *
+	 * The table will be set back into standard mode, and will track the latest
+	 * version of the table.
+	 */
+	async checkoutLatest(): Promise<void> {
+		await this.inner.checkoutLatest();
+	}
 
-  /**
-   * Restore the table to the currently checked out version
-   *
-   * This operation will fail if checkout has not been called previously
-   *
-   * This operation will overwrite the latest version of the table with a
-   * previous version.  Any changes made since the checked out version will
-   * no longer be visible.
-   *
-   * Once the operation concludes the table will no longer be in a checked
-   * out state and the read_consistency_interval, if any, will apply.
-   */
-  async restore(): Promise<void> {
-    await this.inner.restore();
-  }
+	/**
+	 * Restore the table to the currently checked out version
+	 *
+	 * This operation will fail if checkout has not been called previously
+	 *
+	 * This operation will overwrite the latest version of the table with a
+	 * previous version.  Any changes made since the checked out version will
+	 * no longer be visible.
+	 *
+	 * Once the operation concludes the table will no longer be in a checked
+	 * out state and the read_consistency_interval, if any, will apply.
+	 */
+	async restore(): Promise<void> {
+		await this.inner.restore();
+	}
 
-  /** List all indices that have been created with {@link Table.createIndex} */
-  async listIndices(): Promise<IndexConfig[]> {
-    return await this.inner.listIndices();
-  }
+	/** List all indices that have been created with {@link Table.createIndex} */
+	async listIndices(): Promise<IndexConfig[]> {
+		return await this.inner.listIndices();
+	}
 }
